# execdetails: refine cop task execution stats display in plan