Signed-off-by: crazycs520 <crazycs520@gmail.com>

<!-- Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?

As the title said. here is an example:

prepare:

```sql
create table t (a int auto_increment,b int,c int, primary key (a));
insert into t5 (b) values (null),(null),(null),(null),(null),(null);
```

**before this PR**

```sql
explain analyze select count(*) from t5 where a<1200 and b="abc";
+-------------------------+-...-+----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+--------------------------------------------------+-----------+------+
| id                      | ... | execution info                                                                                                                                                                                                 | operator info                                    | memory    | disk |
+-------------------------+-...-+----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+--------------------------------------------------+-----------+------+
| StreamAgg_17            | ... | time:3.08ms, loops:2                                                                                                                                                                                           | funcs:count(Column#7)->Column#5                  | 380 Bytes | N/A  |
| └─IndexReader_18        | ... | time:3.06ms, loops:2, cop_task: {num: 1, max: 3.05ms, proc_keys: 34, rpc_num: 1, rpc_time: 3.02ms, copr_cache_hit_ratio: 0.00}                                                                                 | index:StreamAgg_9                                | 206 Bytes | N/A  |
|   └─StreamAgg_9         | ... | tikv_task:{time:0s, loops:1}, scan_detail: {total_process_keys: 34, total_keys: 35, rocksdb: {delete_skipped_count: 0, key_skipped_count: 34, block: {cache_hit_count: 0, read_count: 0, read_byte: 0 Bytes}}} | funcs:count(1)->Column#7                         | N/A       | N/A  |
|     └─IndexRangeScan_16 | ... | tikv_task:{time:0s, loops:1}, scan_detail: {total_process_keys: 0, total_keys: 0, rocksdb: {delete_skipped_count: 0, key_skipped_count: 0, block: {cache_hit_count: 0, read_count: 0, read_byte: 0 Bytes}}}    | range:[-inf,100), keep order:false, stats:pseudo | N/A       | N/A  |
+-------------------------+-...-+----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+--------------------------------------------------+-----------+------+
```

**this PR**

```sql
+-------------------------+-...-+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+--------------------------------------------------+-----------+------+
| id                      | ... | execution info                                                                                                                                                                                                  | operator info                                    | memory    | disk |
+-------------------------+-...-+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+--------------------------------------------------+-----------+------+
| StreamAgg_17            | ... | time:3.44ms, loops:2                                                                                                                                                                                            | funcs:count(Column#7)->Column#5                  | 380 Bytes | N/A  |
| └─IndexReader_18        | ... | time:3.42ms, loops:2, cop_task: {num: 1, max: 3.3ms, proc_keys: 34, tot_proc: 1ms, tot_wait: 1ms, rpc_num: 1, rpc_time: 3.21ms, copr_cache_hit_ratio: 0.00}                                                     | index:StreamAgg_9                                | 212 Bytes | N/A  |
|   └─StreamAgg_9         | ... | tikv_task:{time:1ms, loops:1}, scan_detail: {total_process_keys: 34, total_keys: 35, rocksdb: {delete_skipped_count: 0, key_skipped_count: 34, block: {cache_hit_count: 0, read_count: 0, read_byte: 0 Bytes}}} | funcs:count(1)->Column#7                         | N/A       | N/A  |
|     └─IndexRangeScan_16 | ... | tikv_task:{time:0s, loops:1}                                                                                                                                                                                    | range:[-inf,100), keep order:false, stats:pseudo | N/A       | N/A  |
+-------------------------+-...-+-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+--------------------------------------------------+-----------+------+
```

As you can see, the `scan_detail` of `└─IndexRangeScan_16` is always 0, so we no need to display it.
In the other hands, 1 cop RPC can only have 1 `scan_detail`, so `scan_detail` information should also appear once. 


### Related changes

- N/A

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

- Unit test
- Manual test (add detailed scripts or steps below)

Side effects

- Performance regression
    - Consumes more CPU

### Release note <!-- bugfixes or new feature need a release note -->

- N/A